[ESLint] Fix LoginScreen
