+++ .eslintrc.json
@@ -39,7 +39,7 @@
         // TODO: Re-enable these accessibility rules
         "jsx-a11y/no-static-element-interactions": "off",
         "jsx-a11y/click-events-have-key-events": "off",
-        "jsx-a11y/interactive-supports-focus": "off"
-
+        "jsx-a11y/interactive-supports-focus": "off",
+        "jsx-a11y/label-has-associated-control": "off"
     }
 }
